Linux Support: make sure that when we get the stack page for thread info we're doing a 64bit not
--HG--
extra : convert_revision : c581921dd601fc72fd2d45b961c7440755b0331c

+++ src/arch/alpha/linux/threadinfo.hh
@@ -57,7 +57,7 @@ class ThreadInfo
          * thread_info struct. So we can get the address by masking off
          * the lower 14 bits.
          */
-        current = tc->readIntReg(TheISA::StackPointerReg) & ~0x3fff;
+        current = tc->readIntReg(TheISA::StackPointerReg) & ~ULL(0x3fff);
         return VPtr<thread_info>(tc, current);
     }
